Move doc gen script to bin/
